+++ symulator/src/StudyEvent.java
@@ -1,8 +1,8 @@
 public class StudyEvent {
     private Subject subject;
-    private int hours;
+    private double hours;
     private int day;
-    StudyEvent(Subject subject, int hours, int day, double motivation) {
+    StudyEvent(Subject subject, double hours, int day, double motivation) {
         this.subject = subject;
         this.hours = hours;
         this.day = day;
